refactor: added vector select box in DataPreviewContent
